oyin_ml/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
chromadb>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
deep-translator>=1.11.0
langdetect>=1.0.9

